Extract ScenarioFixture trait
